Moved unrelated code out of pushdown.go
