+++ sql/analyzer/pushdown.go
@@ -1,9 +1,6 @@
 package analyzer
 
 import (
-	"reflect"
-	"sync"
-
 	"github.com/src-d/go-mysql-server/sql"
 	"github.com/src-d/go-mysql-server/sql/expression"
 	"github.com/src-d/go-mysql-server/sql/plan"
@@ -46,9 +43,7 @@ func pushdown(ctx *sql.Context, a *Analyzer, n sql.Node) (sql.Node, error) {
 	// First step is to find all col exprs and group them by the table they mention.
 	// Even if they appear multiple times, only the first one will be used.
 	a.Log("finding used columns in node")
-	colSpan, _ := ctx.Span("find_pushdown_columns")
 	fieldsByTable := findFieldsByTable(ctx, n)
-	colSpan.Finish()
 
 	a.Log("finding filters in node")
 	filters := findFilters(ctx, n)
@@ -60,46 +55,6 @@ func pushdown(ctx *sql.Context, a *Analyzer, n sql.Node) (sql.Node, error) {
 
 	a.Log("transforming nodes with pushdown of filters, projections and indexes")
 	return transformPushdown(ctx, a, n, filters, indexes, fieldsByTable)
-}
-
-// fixFieldIndexesOnExpressions executes fixFieldIndexes on a list of exprs.
-func fixFieldIndexesOnExpressions(schema sql.Schema, expressions ...sql.Expression) ([]sql.Expression, error) {
-	var result = make([]sql.Expression, len(expressions))
-	for i, e := range expressions {
-		var err error
-		result[i], err = fixFieldIndexes(schema, e)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return result, nil
-}
-
-// fixFieldIndexes transforms the given expression setting correct indexes
-// for GetField expressions according to the schema of the row in the table
-// and not the one where the filter came from.
-func fixFieldIndexes(schema sql.Schema, exp sql.Expression) (sql.Expression, error) {
-	return expression.TransformUp(exp, func(e sql.Expression) (sql.Expression, error) {
-		switch e := e.(type) {
-		case *expression.GetField:
-			// we need to rewrite the indexes for the table row
-			for i, col := range schema {
-				if e.Name() == col.Name && e.Table() == col.Source {
-					return expression.NewGetFieldWithTable(
-						i,
-						e.Type(),
-						e.Table(),
-						e.Name(),
-						e.IsNullable(),
-					), nil
-				}
-			}
-
-			return nil, ErrFieldMissing.New(e.Name())
-		}
-
-		return e, nil
-	})
 }
 
 func findFieldsByTable(ctx *sql.Context, n sql.Node) map[string][]string {
@@ -197,69 +152,6 @@ func transformPushdown(
 	return node, nil
 }
 
-// Transforms the expressions in the Node given, fixing the field indexes.
-func fixFieldIndexesForExpressions(node sql.Node) (sql.Node, error) {
-	if _, ok := node.(sql.Expressioner); !ok {
-		return node, nil
-	}
-
-	var schemas []sql.Schema
-	for _, child := range node.Children() {
-		schemas = append(schemas, child.Schema())
-	}
-
-	if len(schemas) < 1 {
-		return node, nil
-	}
-
-	n, err := plan.TransformExpressions(node, func(e sql.Expression) (sql.Expression, error) {
-		for _, schema := range schemas {
-			fixed, err := fixFieldIndexes(schema, e)
-			if err == nil {
-				return fixed, nil
-			}
-
-			if ErrFieldMissing.Is(err) {
-				continue
-			}
-
-			return nil, err
-		}
-
-		return e, nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	switch j := n.(type) {
-	case *plan.InnerJoin:
-		cond, err := fixFieldIndexes(j.Schema(), j.Cond)
-		if err != nil {
-			return nil, err
-		}
-
-		n = plan.NewInnerJoin(j.Left, j.Right, cond)
-	case *plan.RightJoin:
-		cond, err := fixFieldIndexes(j.Schema(), j.Cond)
-		if err != nil {
-			return nil, err
-		}
-
-		n = plan.NewRightJoin(j.Left, j.Right, cond)
-	case *plan.LeftJoin:
-		cond, err := fixFieldIndexes(j.Schema(), j.Cond)
-		if err != nil {
-			return nil, err
-		}
-
-		n = plan.NewLeftJoin(j.Left, j.Right, cond)
-	}
-
-	return n, nil
-}
-
 func pushdownTable(
 	a *Analyzer,
 	node *plan.ResolvedTable,
@@ -334,72 +226,4 @@ func pushdownFilter(
 	)
 
 	return plan.NewFilter(expression.JoinAnd(unhandled...), node.Child), nil
-}
-
-type Releaser struct {
-	Child   sql.Node
-	Release func()
-}
-
-func (r *Releaser) Resolved() bool {
-	return r.Child.Resolved()
-}
-
-func (r *Releaser) Children() []sql.Node {
-	return []sql.Node{r.Child}
-}
-
-func (r *Releaser) RowIter(ctx *sql.Context) (sql.RowIter, error) {
-	iter, err := r.Child.RowIter(ctx)
-	if err != nil {
-		r.Release()
-		return nil, err
-	}
-
-	return &releaseIter{child: iter, release: r.Release}, nil
-}
-
-func (r *Releaser) Schema() sql.Schema {
-	return r.Child.Schema()
-}
-
-func (r *Releaser) WithChildren(children ...sql.Node) (sql.Node, error) {
-	if len(children) != 1 {
-		return nil, sql.ErrInvalidChildrenNumber.New(r, len(children), 1)
-	}
-	return &Releaser{children[0], r.Release}, nil
-}
-
-func (r *Releaser) String() string {
-	return r.Child.String()
-}
-
-func (r *Releaser) Equal(n sql.Node) bool {
-	if r2, ok := n.(*Releaser); ok {
-		return reflect.DeepEqual(r.Child, r2.Child)
-	}
-	return false
-}
-
-type releaseIter struct {
-	child   sql.RowIter
-	release func()
-	once    sync.Once
-}
-
-func (i *releaseIter) Next() (sql.Row, error) {
-	row, err := i.child.Next()
-	if err != nil {
-		_ = i.Close()
-		return nil, err
-	}
-	return row, nil
-}
-
-func (i *releaseIter) Close() (err error) {
-	i.once.Do(i.release)
-	if i.child != nil {
-		err = i.child.Close()
-	}
-	return err
 }
